experiments/cl_cr3/pipeline/requirements.txt
# Pseudo requirements file
numpy
pandas
scikit-learn
matplotlib
tensorflow
torch
h5py
pyyaml
transformers
pynvml
openai
accelerate
seaborn
torchvision 
torchaudio 
datasets